layers: New sparse map for image subresource data
Implements an range map and subresource adapters suitable for
synchronization work, integrated as a replacement for the
ImageSubresourceLayout map to test for robustness, correctness, and
performance.

As is, is slower than the previous ImageSubresourceMap, before tuning,
but new iterator should all for lambda (and new/free) free image
subresource tracking.

Change-Id: Ie7146d94047df227967357f958135d5d2e3b6aff